+++ crates/pallets/starknet/src/lib.rs
@@ -85,8 +85,8 @@ use mp_sequencer_address::{InherentError, InherentType, DEFAULT_SEQUENCER_ADDRES
 use mp_storage::{StarknetStorageSchemaVersion, PALLET_STARKNET_SCHEMA};
 use mp_transactions::execution::Execute;
 use mp_transactions::{
-    DeclareTransaction, DeployAccountTransaction, HandleL1MessageTransaction, InvokeTransaction, Transaction,
-    UserOrL1HandlerTransaction, UserTransaction,
+    DeclareTransaction, DeployAccountTransaction, DeployTransaction, HandleL1MessageTransaction, InvokeTransaction,
+    Transaction, UserOrL1HandlerTransaction, UserTransaction,
 };
 use sp_runtime::traits::UniqueSaturatedInto;
 use sp_runtime::DigestItem;
@@ -787,6 +787,55 @@ pub mod pallet {
                 Transaction::L1Handler(input_transaction),
                 tx_execution_infos.revert_error,
             );
+
+            Ok(())
+        }
+
+        /// DeployTransaction, type of transaction used before Starknet v0.10.1.
+        /// Before you can calculate the transaction hash, get the deployed contract address.
+        /// # Arguments
+        ///
+        /// * `origin` - The origin of the transaction.
+        /// * `transaction` - The Starknet transaction.
+        ///
+        ///  # Returns
+        ///
+        /// * `DispatchResult` - The result of the transaction.
+        #[pallet::call_index(5)]
+        #[pallet::weight({0})]
+        pub fn deploy(origin: OriginFor<T>, transaction: DeployTransaction) -> DispatchResult {
+            // This ensures that the function can only be called via unsigned transaction.
+            ensure_none(origin)?;
+
+            let input_transaction = transaction;
+            let chain_id = T::ChainId::get();
+            let transaction = input_transaction.into_executable::<T::SystemHash>(chain_id, false);
+
+            // Check if contract is deployed
+            ensure!(
+                !ContractClassHashes::<T>::contains_key(transaction.contract_address),
+                Error::<T>::AccountAlreadyDeployed
+            );
+
+            // Execute
+            let tx_execution_infos = transaction
+                .execute(
+                    &mut BlockifierStateAdapter::<T>::default(),
+                    &Self::get_block_context(),
+                    &RuntimeExecutionConfigBuilder::new::<T>().build(),
+                )
+                .map_err(|e| {
+                    log::error!("failed to deploy transaction: {:?}", e);
+                    Error::<T>::TransactionExecutionFailed
+                })?;
+
+            let tx_hash = transaction.tx_hash;
+            Self::emit_and_store_tx_and_fees_events(
+                tx_hash,
+                &tx_execution_infos.execute_call_info,
+                &tx_execution_infos.fee_transfer_call_info,
+            );
+            Self::store_transaction(tx_hash, Transaction::Deploy(input_transaction), tx_execution_infos.revert_error);
 
             Ok(())
         }