+++ CHANGELOG.md
@@ -19,6 +19,7 @@ git # Madara Changelog
 - feat: add transparent representation to `Felt252Wrapper`
 - feat(rpc/trace_api): add `trace_block_transaction`
 - chore(db): changed the way hashes are encoded
+- feat: add DeployTransaction implementation
 
 ## v0.7.0